fixed auto cast/coversion warning

+++ Example/ExampleObjectiveC/PIONotificationViewController.m
@@ -99,7 +99,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    id <NSFetchedResultsSectionInfo> sectionInfo = self.fetchedResultsController.sections[section];
+    id <NSFetchedResultsSectionInfo> sectionInfo = self.fetchedResultsController.sections[(NSUInteger) section];
     return sectionInfo.numberOfObjects;
 }
 
@@ -130,7 +130,7 @@
 - (void)configureCell:(UITableViewCell *)cell withEvent:(EventViaNotification *)event {
     enum PredictIOEventType eventType = (enum PredictIOEventType) event.type.integerValue;
     if (eventType == TransportMode) {
-        cell.textLabel.text = self.transportationModeLabels[event.mode.integerValue];
+        cell.textLabel.text = self.transportationModeLabels[(NSUInteger) event.mode.integerValue];
     } else {
         cell.textLabel.text = self.labels[eventType];
     }
